Support upload videos. #128

+++ src/tasks/media-worker.js
@@ -170,21 +170,14 @@ module.exports = (job, done) => {
           throw new errors.Http403();
         }
         // Do validations for each the file type.
-        switch (file.type) {
-          case FileType.annotationImage:
-          case FileType.annotationZIP:
-          case FileType.annotationCSV:
-            if (!cameraLocation) {
-              throw new errors.Http404('Camera location is not found.');
-            }
-            if (
-              !cameraLocation.studyArea ||
-              cameraLocation.studyArea.state !== StudyAreaState.active
-            ) {
-              throw new errors.Http404('Study area is not found.');
-            }
-            break;
-          default:
+        if (!cameraLocation) {
+          throw new errors.Http404('Camera location is not found.');
+        }
+        if (
+          !cameraLocation.studyArea ||
+          cameraLocation.studyArea.state !== StudyAreaState.active
+        ) {
+          throw new errors.Http404('Study area is not found.');
         }
       },
     )
@@ -211,12 +204,27 @@ module.exports = (job, done) => {
         const file = fs.createWriteStream(tempFile.name);
 
         file.on('close', () => {
-          extract(tempFile.name, { dir: tempDir.name }, error => {
-            if (error) {
-              return reject(error);
-            }
-            resolve();
-          });
+          extract(
+            tempFile.name,
+            {
+              dir: tempDir.name,
+              onEntry: (entry, zipfile) => {
+                zipfile.once('end', () => {
+                  fs.utimesSync(
+                    path.join(tempDir.name, entry.fileName),
+                    entry.getLastModDate(),
+                    entry.getLastModDate(),
+                  );
+                });
+              },
+            },
+            error => {
+              if (error) {
+                return reject(error);
+              }
+              resolve();
+            },
+          );
         });
         s3.getObject({
           Bucket: config.s3.bucket,
@@ -249,25 +257,34 @@ module.exports = (job, done) => {
                 if (
                   FileExtensionName.annotationImage.indexOf(fileExtensionName) <
                     0 &&
-                  fileExtensionName.annotationCsv.indexOf(fileExtensionName) < 0
+                  FileExtensionName.annotationVideo.indexOf(fileExtensionName) <
+                    0 &&
+                  FileExtensionName.annotationCsv.indexOf(fileExtensionName) < 0
                 ) {
                   // This is not allow files.
                   fs.unlinkSync(path.join(tempDir.name, filename));
                   return;
                 }
-                const content = fs.readFileSync(
-                  path.join(tempDir.name, filename),
-                );
+                let source = path.join(tempDir.name, filename);
+
+                // Fix file.type
                 if (
                   FileExtensionName.annotationCsv.indexOf(fileExtensionName) >=
                   0
                 ) {
-                  // Fix file.type
+                  // The binary content will be used for convert to annotations.
+                  source = fs.readFileSync(source);
+                  file.content = source;
                   file.type = FileType.annotationCSV;
-                  file.content = content; // The binary content will be used for convert to annotations.
+                } else if (
+                  FileExtensionName.annotationVideo.indexOf(
+                    fileExtensionName,
+                  ) >= 0
+                ) {
+                  file.type = FileType.annotationVideo;
                 }
 
-                return file.saveWithContent(content).then(() => {
+                return file.saveWithContent(source).then(() => {
                   fs.unlinkSync(path.join(tempDir.name, filename));
                   if (
                     file.type === FileType.annotationImage &&
@@ -343,6 +360,7 @@ module.exports = (job, done) => {
       }
 
       // _file.type === FileType.annotationImage
+      // _file.type === FileType.annotationVideo
       // _file.type === FileType.annotationZIP
       const csvFileIndex = files.findIndex(
         x => x.type === FileType.annotationCSV,
@@ -451,7 +469,11 @@ module.exports = (job, done) => {
        */
       const tasks = [];
 
-      if (_file.type === FileType.annotationImage) {
+      if (
+        [FileType.annotationImage, FileType.annotationVideo].indexOf(
+          _file.type,
+        ) >= 0
+      ) {
         annotations.forEach(annotation => {
           const duplicateAnnotation = duplicateAnnotations.find(
             x =>
